feat : CustomError 초기 구현

See merge request s11-final/S11P31A603!18

+++ backend/src/main/java/com/ssafy/fittapet/backend/application/service/map/MapServiceImpl.java
@@ -1,8 +1,10 @@
 package com.ssafy.fittapet.backend.application.service.map;
 
+import com.ssafy.fittapet.backend.common.exception.CustomException;
 import com.ssafy.fittapet.backend.common.util.EnteringCodeUtil;
 import com.ssafy.fittapet.backend.common.validator.GuildValidator;
 import com.ssafy.fittapet.backend.common.validator.MapValidator;
+import com.ssafy.fittapet.backend.domain.dto.guild.GuildJoinRequest;
 import com.ssafy.fittapet.backend.domain.dto.guild.GuildRequest;
 import com.ssafy.fittapet.backend.domain.dto.map.MapResponse;
 import com.ssafy.fittapet.backend.domain.entity.Guild;
@@ -17,6 +19,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
+import static com.ssafy.fittapet.backend.common.constant.error_code.GuildErrorCode.*;
 
 @Service
 @RequiredArgsConstructor
@@ -40,14 +44,14 @@ public class MapServiceImpl implements MapService {
     }
 
     @Override
-    public void createGuild(GuildRequest guildRequest) {
+    public void createGuild(GuildRequest guildRequest) throws CustomException {
         // 1. todo : 로그인한 유저 id
         Long userId = 1L;
         User user = User.builder().id(userId).build();
         // 2. position 유효 검사
-        if(!mapValidator.isAblePosition(userId, guildRequest.getGuildPosition())) return;
+        if(!mapValidator.isAblePosition(userId, guildRequest.getGuildPosition())) throw new CustomException(NOT_AVAILABLE_POSITION);
         // 3. 그룹 이름 유효 검사?
-        if(!guildValidator.isNameUnique(guildRequest.getGuildName())) return;
+        if(!guildValidator.isNameUnique(guildRequest.getGuildName())) throw new CustomException(DUPLICATED_NAME);
         // 4. 그룹 생성
         Guild guild = Guild.builder().
                 guildLeader(user).
@@ -63,22 +67,26 @@ public class MapServiceImpl implements MapService {
     }
 
     @Override
-    public Boolean joinGuild(String enteringCode, Long guildPosition) {
-        try {
+    public Boolean joinGuild(GuildJoinRequest guildJoinRequest) throws Exception {
+
             // todo : 요청자 정보 받아오기
             User user = User.builder().id(1L).build();
+
+            String enteringCode = guildJoinRequest.getEnteringCode();
+            Long guildPosition = guildJoinRequest.getGuildPosition();
 
             // 초대 코드 기간이 유효하면 guildId, 유효하지 않으면 -1 반환
             Long guildId = EnteringCodeUtil.isCodeValid(enteringCode);
             if(guildId == -1) return false;
-            // 그룹 없으면 가입 불가
-            Guild guild = guildValidator.isExist(guildId).orElseThrow();
+            // 길드 없으면 가입 불가
+            Guild guild = guildValidator.isExist(guildId).orElseThrow(() -> new CustomException(NO_GUILD));
             // 이미 가입되어 있으면 가입 불가
-            if(mapValidator.isAlreadyJoined(user.getId(), guildId)) return false;
+            if(mapValidator.isAlreadyJoined(user.getId(), guildId)) throw new CustomException(ALREADY_JOIN);
             // 그룹 인원 수가 6명보다 많으면 가입 불가
-            if(!mapValidator.isUnder6(guildId)) return false;
+            if(!mapValidator.isUnder6(guildId)) throw new CustomException(FULL_GUILD);
             // 위치가 유효하지 않으면 가입 불가
-            if(!mapValidator.isAblePosition(user.getId(), guildPosition)) return false;
+            if(!mapValidator.isAblePosition(user.getId(), guildPosition)) throw new CustomException(NOT_AVAILABLE_POSITION);
+
             // 조건 다 만족하면 가입
             mapRepository.save(Map.builder().
                     user(user).
@@ -87,24 +95,21 @@ public class MapServiceImpl implements MapService {
                     build());
 
             return true;
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
+
     }
 
     @Override
-    public void leaveGuild(Long guildId) {
+    public void leaveGuild(Long guildId) throws CustomException {
         // map에서 삭제
         // todo : 요청자 정보 가져오기
         User user = User.builder().id(1L).build();
-        if(!mapValidator.isAlreadyJoined(user.getId(), guildId)) return;
-        if(!guildValidator.isExist(guildId).equals(null)) return;
+        if(!mapValidator.isAlreadyJoined(user.getId(), guildId)) throw new CustomException(NOT_GUILD_MEMBER);
+        if(!guildValidator.isExist(guildId).equals(null)) throw new CustomException(NO_GUILD);
 
         mapRepository.deleteByGuildIdAndUserId(guildId, user.getId());
 
         // userQuestStatus에서 삭제
         GuildQuest guildQuest = guildQuestRepository.findByGuildId(guildId);
-        if(guildQuest == null) return;
-        userQuestStatusRepository.deleteByUserIdAndGuildQuestId(user.getId(), guildQuest.getId());
+        if(guildQuest != null) userQuestStatusRepository.deleteByUserIdAndGuildQuestId(user.getId(), guildQuest.getId());
     }
 }
